Resolve cyclic commit
